# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": " uala twitter-like API",
        "description": "Documentación de la API generada con Swagger/OpenAPI",
        "contact": {
            "email": "estebandiaczun@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/users": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Crear un nuevo usuario",
                "description": "Crea un nueevo usuario y devuelve su información.",
                "operationId": "ff67e6b5ba3334edbe90020b702bd441",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username"
                                ],
                                "properties": {
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "display_name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "550e8400-e29b-41d4-a716-446655440000"
                                                },
                                                "username": {
                                                    "type": "string",
                                                    "example": "johndoe"
                                                },
                                                "display_name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "username": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The username has already been taken."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{id}": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Obtener información de un usuario",
                "description": "Este endpoint permite obtener la información de un usuario especificado mediante su ID. Se requiere pasar un encabezado `X-User-ID` con el ID del usuario autenticado como token.",
                "operationId": "getUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del usuario que se desea obtener",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-User-ID",
                        "in": "header",
                        "description": "ID del usuario autenticado, enviado como encabezado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del usuario obtenida exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "123e4567-e89b-12d3-a456-426614174000"
                                        },
                                        "username": {
                                            "type": "string",
                                            "example": "johndoe"
                                        },
                                        "display_name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "stats": {
                                            "properties": {
                                                "followers_count": {
                                                    "type": "integer",
                                                    "example": 150
                                                },
                                                "following_count": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Solicitud inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid request."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{targetUserId}/follow": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "Seguir a un usuario",
                "description": "Permite que el usuario autenticado siga a otro usuario. El sistema actualiza los contadores de seguidores tanto en la base de datos como en caché.",
                "operationId": "followUser",
                "parameters": [
                    {
                        "name": "targetUserId",
                        "in": "path",
                        "description": "ID del usuario a seguir",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "X-User-ID",
                        "in": "header",
                        "description": "ID del usuario autenticado, enviado como encabezado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario seguido exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User followed successfully"
                                        },
                                        "stats": {
                                            "properties": {
                                                "followers_count": {
                                                    "type": "integer",
                                                    "example": 42
                                                },
                                                "following_count": {
                                                    "type": "integer",
                                                    "example": 23
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "400": {
                        "description": "No se puede seguir al mismo usuario"
                    }
                },
                "security": [
                    {
                        "X-User-ID": []
                    }
                ]
            }
        },
        "/api/v1/users/{targetUserId}/unfollow": {
            "delete": {
                "tags": [
                    "User Management"
                ],
                "summary": "Dejar de seguir a un usuario",
                "description": "Permite que el usuario autenticado deje de seguir a otro usuario. El sistema actualiza los contadores de seguidores tanto en la base de datos como en caché.",
                "operationId": "unfollowUser",
                "parameters": [
                    {
                        "name": "X-User-ID",
                        "in": "header",
                        "description": "ID del usuario autenticado, enviado como encabezado",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "targetUserId",
                        "in": "path",
                        "description": "ID del usuario a dejar de seguir",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario dejado de seguir exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User unfollowed successfully"
                                        },
                                        "stats": {
                                            "properties": {
                                                "followers_count": {
                                                    "type": "integer",
                                                    "example": 41
                                                },
                                                "following_count": {
                                                    "type": "integer",
                                                    "example": 22
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuario no encontrado"
                    },
                    "400": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Cannot unfollow yourself."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "User is not being followed."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "X-User-ID": []
                    }
                ]
            }
        },
...
    },
...
}